+++ autosubmit/platforms/wrappers/wrapper_factory.py
@@ -21,7 +21,7 @@ from autosubmit.platforms.wrappers.wrapper_builder import (
     WrapperDirector, PythonVerticalWrapperBuilder, PythonHorizontalWrapperBuilder,
     PythonHorizontalVerticalWrapperBuilder, PythonVerticalHorizontalWrapperBuilder, BashHorizontalWrapperBuilder,
     BashVerticalWrapperBuilder, SrunHorizontalWrapperBuilder, SrunVerticalHorizontalWrapperBuilder,
-    FluxWrapperBuilder
+    FluxWrapperBuilder, HyperQueueWrapperBuilder
 )
 
 
@@ -214,6 +214,8 @@ class SlurmWrapperFactory(WrapperFactory):
     def delegated_wrapper(self, **kwargs):
         if kwargs["method"] == 'flux':
             return FluxWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'hyperqueue':
+            return HyperQueueWrapperBuilder(**kwargs)
         else:
             raise NotImplementedError(self.exception_method(kwargs["method"]))  # pragma: no cover